Make clang-tidy happy

+++ src/fheroes2/world/world_loadmap.cpp
@@ -692,7 +692,7 @@ bool World::loadResurrectionMap( const std::string & filename )
             if ( object.group == Maps::ObjectGroup::KINGDOM_TOWNS ) {
                 const uint32_t race = townObjects[object.index].metadata[0];
 
-                Castle * castle = new Castle( static_cast<int32_t>( tileId ) % width, static_cast<int32_t>( tileId ) / width, race );
+                Castle * castle = new Castle( static_cast<int32_t>( tileId ) % width, static_cast<int32_t>( tileId ) / width, static_cast<int>( race ) );
                 const uint8_t color = Maps::getTownColorIndex( map, tileId, object.id );
                 castle->SetColor( 1 << color );
